anderson-work/requirements.txt
livekit-agents>=0.11.2
livekit-plugins-openai>=0.10.6
livekit-plugins-deepgram>=0.6.9
livekit-plugins-silero>=0.7.3
python-dotenv~=1.0
tiktoken=0.8.0
langchain=0.3.7
langchain-openai=0.2.9
openai=1.57.5
supabase=2.10.0